fixed some errors with updating contacts

+++ PageContactsTable.php
@@ -12,16 +12,21 @@ class PageContactsTable
     {
         ob_start();
         echo "<h1>Contacts List</h1>";
-        
+    
         $this->handleRequest();
-
-        $this->renderContactsTable();
-
-        $this->renderCreateForm();
-
+    
+        if (isset($_GET['action']) && $_GET['action'] === 'edit' && isset($_GET['id'])) {
+            $contact = new DBContact($_GET['id']);
+            $this->renderEditForm($contact);
+        } else {
+            $this->renderContactsTable();
+            $this->renderCreateForm();
+        }
+    
         $html = ob_get_clean();
         return $html;
     }
+    
 
     private function handleRequest()
     {
@@ -34,18 +39,33 @@ class PageContactsTable
                 $contact->phone_number = $_POST['phone_number'];
                 $contact->email = $_POST['email'];
                 $contact->save();
+    
+                header("Location: index.php");
+                exit();
             } elseif ($action === 'update') {
-                $contact = new DBContact($_POST['id']);
-                $contact->name = $_POST['name'];
-                $contact->phone_number = $_POST['phone_number'];
-                $contact->email = $_POST['email'];
-                $contact->save();
+                if (!empty($_POST['id'])) {
+                    $contact = new DBContact($_POST['id']);
+                    $contact->name = $_POST['name'];
+                    $contact->phone_number = $_POST['phone_number'];
+                    $contact->email = $_POST['email'];
+                    $contact->save();
+    
+                    header("Location: index.php");
+                    exit();
+                }
             }
         } elseif (isset($_GET['action']) && $_GET['action'] === 'delete') {
-            $contact = new DBContact($_GET['id']);
-            $contact->delete();
+            if (!empty($_GET['id'])) {
+                $contact = new DBContact($_GET['id']);
+                $contact->delete();
+    
+                header("Location: index.php");
+                exit();
+            }
         }
     }
+    
+    
 
     private function renderContactsTable()
     {
@@ -74,7 +94,7 @@ class PageContactsTable
             $tr->contained .= $tdEmail->draw();
 
             $tdActions = new TableTD();
-            $tdActions->contained = "<a href='?action=update&id={$contact->id}'>Edit</a> | "
+            $tdActions->contained = "<a href='?action=edit&id={$contact->id}'>Edit</a> | "
                 . "<a href='?action=delete&id={$contact->id}' onclick='return confirm(\"Are you sure?\")'>Delete</a>";
             $tr->contained .= $tdActions->draw();
 
@@ -83,6 +103,7 @@ class PageContactsTable
 
         echo "</table>";
     }
+
 
     private function renderCreateForm()
     {
@@ -111,4 +132,37 @@ class PageContactsTable
         echo "<input type='submit' value='Add Contact'>";
         echo "</form>";
     }
+
+    private function renderEditForm($contact)
+    {
+        echo "<h2>Edit Contact</h2>";
+        echo "<form method='post'>";
+        echo "<input type='hidden' name='action' value='update'>";
+        echo "<input type='hidden' name='id' value='{$contact->id}'>";
+
+        echo "<label>Name:</label>";
+        $nameInput = new Input();
+        $nameInput->attributes->name = "name";
+        $nameInput->attributes->value = $contact->name;
+        echo $nameInput->draw();
+        echo "<br>";
+
+        echo "<label>Phone:</label>";
+        $phoneInput = new Input();
+        $phoneInput->attributes->name = "phone_number";
+        $phoneInput->attributes->value = $contact->phone_number;
+        echo $phoneInput->draw();
+        echo "<br>";
+
+        echo "<label>Email:</label>";
+        $emailInput = new Input();
+        $emailInput->attributes->name = "email";
+        $emailInput->attributes->value = $contact->email;
+        echo $emailInput->draw();
+        echo "<br>";
+
+        echo "<input type='submit' value='Update Contact'>";
+        echo "</form>";
+    }
+
 }
